Improve nesting logic so it works in jQuery < 1.9
After improving the test suite to run against all supported versions of
jQuery each time, I discovered that the LIFO ordering test was failing
in all older versions. This turned out to be due to a different in the
way prefilters and transports are processed in jQuery >= 1.9.

In newer jQuery, the first prefilter with change the dataType, then its
transport will activate and handle it. Then the next one will step up
and do the same. In older versions, however, all of the prefilters will
apply first, and then only the last transport will activate, which
prevent them from nesting properly.

Details
-------

* Move setting the flag that determines whether a middleware is done
  with things to the transport.
* Set an additional flag in the prefilter to indicate that some
  middleware is already handling the request, and unset it in the
  transport.