Change name option to namespace and update getBy methods
`name` was never a great name for the option, so I have changed it to
`namespace`. Also, since jQuery essentially forces dataTypes to be
case-insensitive, I enforced that in the `getByNamespace` and new
`getByDataType` methods. Additionally, I made the activeness-related methods
(`activate`, `deactivate` and `isActive`) use a truly private state variable
rather than leaving it exposed.

I also bumped the version up to 0.3.0 to indicate the change in API.